requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic[email]==2.12.5

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pyjwt==2.8.0
python-dotenv==1.0.1
pydantic-settings==2.1.0

# OCR Dependencies
python-magic==0.4.27
pytesseract==0.3.10
pdfplumber==0.11.0
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
striprtf==0.0.26
beautifulsoup4==4.12.3
EbookLib==0.18
Pillow==10.2.0
lxml==5.1.0
odfpy==1.4.1
markdown-it-py==3.0.0

# Async & Background Tasks
celery==5.3.6
redis==5.0.1

# HTTP & Networking
httpx==0.26.0

# Monitoring & Logging
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
